First working emulation (not complete yet)
Can load Chip-8 program and run it. Emulator struct is the emulator
itself, than needs to be created with a display instance that implements
Display. Added DisplayTermbox that could be used as default Display.

Errors are not propagated properly and not managed. Timers are missing.
Collision checks when drawing and boundary checks are not done. Not
checked when PC goes out of memory bounds.

Signed-off-by: Deividas Puplauskas <ngivingup@gmail.com>